Update Question.cs
Changed the regex to not fail on names that contain hyphens.  It may even be appropriate to drop the domain name regex entirely because I would expect the server to answer the question of whether the domain name is valid.  Note that the domain name will still fail validation on names with just one part (e.g., server1, server2, etc.)